con-rs: extract PraosState sans-IO state machine
Pulls every consensus-shaped field out of net-node's `PraosConsensus`
into a new `con_rs::praos::PraosState`.  The wrapper that remains in
net-node is purely I/O glue: it translates `NetworkEvent` /
`LedgerOutcome` / wire-type args into logical args, calls into the
state machine, and dispatches the returned `Vec<PraosEffect>` to the
network-command channel and validator actor.

Moved into `PraosState`:
- chain_tree, adopted_tip_hash
- block_cache (with opaque header/body bytes; `CachedBlock` no longer
  carries `WrappedHeader` / `BlockBody`, sidestepping a wire-format
  dep on con-rs)
- validated, self_produced
- peer_chains, orphan_cooldown
- in_flight (fetch tracking), in_flight_validation
- queued_validator_tip, last_validated_tip
- security_param_k, node_id

Algorithms moved with the state:
- walk_ancestors_hybrid, select_chain_once, try_switch_to (pure)
- evaluate_and_fetch, execute_switch, submit_for_validation,
  issue_fetch, retry_select_chain (now emit `PraosEffect`s rather
  than calling I/O directly)
- on_block_received, on_block_applied, on_block_rolled_back,
  on_block_apply_failed, on_tip_advanced, on_peer_*, on_block_fetch_failed
- register_self_produced

`PraosEffect` covers the I/O surface: FetchBlockRange, ReIntersect,
InjectBlock / InjectRollback (chain-store side), ValidatorApply /
ValidatorRollback (ledger side).

To support the simulator's determinism contract, all collections in
con-rs are `BTreeMap` / `BTreeSet` (peer_chains, block_cache,
orphan_cooldown, in_flight, validated, in_flight_validation,
self_produced).  `Point` gains `PartialOrd` / `Ord` so it can key
`BTreeSet<Point>` for `self_produced`.

Time is injected: methods that need wall-clock take `now: Instant`
(orphan-cooldown expiry, in-flight TTL, etc.).  Sim-rs's adapter will
pass synthetic time; net-node passes `Instant::now()` at the call
boundary.

Net-node delta:
- `consensus/praos/mod.rs` shrinks from a 2200-line file backed by
  `selection.rs` / `validation.rs` / `fetching.rs` (~1700 lines
  combined) to a single 340-line wrapper.  Helper files deleted.
- Public API on `PraosConsensus` is unchanged.  Tests retained,
  poking through `consensus.state.<field>`.
- `dispatch(Vec<PraosEffect>)` translates each effect: rehydrates
  `header: Vec<u8>` / `body: Vec<u8>` into `WrappedHeader` /
  `BlockBody` for `InjectBlock` and `ValidatorApply`.

Verified: 58 con-rs tests pass, 154 net-node tests pass (baseline
154 — no regression), 528 net-rs workspace tests pass, cargo clippy
--all-targets shows 11 warnings in net-rs (baseline) and 0 in con-rs.
